refactor(testkit): one shared test-support crate for tracing and HTTP mocks
AlwaysOnSubscriber existed as five copy-pasted test_support.rs files in
two silently divergent designs (thread-local guard vs process-global
OnceLock), and the raw-TCP mock HTTP server as nine copies across the
provider, package, and setup test suites - about 650 duplicated lines
drifting independently.

crates/leviath-testkit (publish = false, dev-dependency only) now holds
one subscriber offering both install shapes, the panic-hook serializer,
and a mock-server family built on one raw primitive. Each crate's
test_support.rs shrinks to re-exports so established
crate::test_support::* paths keep resolving.

The coverage gate excludes leviath-testkit for the same reason it
excludes xtask: every line here executes inside other packages' gated
suites, and self-gating scaffolding at 100% forces tests-of-test-helpers
with no defect-finding power. parse_workspace_packages records that
decision with its test.
