+++ crates/leviath-runtime/src/test_support.rs
@@ -1,79 +1,4 @@
 #![cfg(test)]
-//! Shared test-only tracing-coverage helper — exactly ONE copy per crate on
-//! purpose. `tracing::subscriber::set_global_default` only succeeds once per
-//! test binary, so a single shared subscriber keeps every trait method
-//! reachable as the active subscriber.
+//! Crate-local names for the shared helpers in `leviath-testkit`.
 
-pub(crate) struct AlwaysOnSubscriber;
-
-impl tracing::Subscriber for AlwaysOnSubscriber {
-    fn enabled(&self, _metadata: &tracing::Metadata<'_>) -> bool {
-        true
-    }
-    fn register_callsite(
-        &self,
-        _metadata: &'static tracing::Metadata<'static>,
-    ) -> tracing::subscriber::Interest {
-        tracing::subscriber::Interest::always()
-    }
-    fn new_span(&self, _span: &tracing::span::Attributes<'_>) -> tracing::span::Id {
-        tracing::span::Id::from_u64(1)
-    }
-    fn record(&self, _span: &tracing::span::Id, _values: &tracing::span::Record<'_>) {}
-    fn record_follows_from(&self, _span: &tracing::span::Id, _follows: &tracing::span::Id) {}
-    fn event(&self, event: &tracing::Event<'_>) {
-        let _ = self.enabled(event.metadata());
-    }
-    fn enter(&self, _span: &tracing::span::Id) {}
-    fn exit(&self, _span: &tracing::span::Id) {}
-    fn max_level_hint(&self) -> Option<tracing::metadata::LevelFilter> {
-        Some(tracing::metadata::LevelFilter::TRACE)
-    }
-}
-
-/// Serializes every test that swaps the global panic hook. Without it two such
-/// tests interleave under the parallel test runner: one test's `set_hook`
-/// replaces the other's silencing closure before the other's panic fires, so
-/// that closure never runs and shows as uncovered. Hold this across each test's
-/// set-hook → panic → restore-hook sequence. (Only ever held across synchronous
-/// work.) One lock per test binary, so it must live here rather than in any one
-/// module's test block.
-pub(crate) static PANIC_HOOK_LOCK: std::sync::Mutex<()> = std::sync::Mutex::new(());
-
-/// Run `f` with the panic hook silenced, restoring the previous hook after.
-/// Returns whatever `f` returns; `f` is expected to swallow the panic itself
-/// (e.g. via `catch_unwind`).
-pub(crate) fn with_silenced_panics<T>(f: impl FnOnce() -> T) -> T {
-    let _guard = PANIC_HOOK_LOCK.lock().expect("panic-hook lock poisoned");
-    let prev = std::panic::take_hook();
-    std::panic::set_hook(Box::new(|_| {}));
-    let out = f();
-    std::panic::set_hook(prev);
-    out
-}
-
-pub(crate) fn with_tracing<T>(f: impl FnOnce() -> T) -> T {
-    static INSTALLED: std::sync::OnceLock<()> = std::sync::OnceLock::new();
-    INSTALLED.get_or_init(|| {
-        let _ = tracing::subscriber::set_global_default(AlwaysOnSubscriber);
-        tracing::callsite::rebuild_interest_cache();
-    });
-    f()
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn always_on_subscriber_span_methods_are_all_no_ops() {
-        with_tracing(|| {
-            let span = tracing::info_span!("test-span", field = tracing::field::Empty);
-            span.record("field", 1);
-            let other = tracing::info_span!("other-span");
-            span.follows_from(&other);
-            let _enter = span.enter();
-            tracing::info!(parent: &span, "inside span");
-        });
-    }
-}
+pub(crate) use leviath_testkit::{PANIC_HOOK_LOCK, with_silenced_panics, with_tracing};
